Conform the Equatable protocol from all the scheme-related models
